Got rid of old dashboard components
